requirements.txt
streamlit
python-dotenv
openai
requests
beautifulsoup4